Refactor components, Update styling and Readme
