+++ src/styles/base/_base.scss
@@ -16,8 +16,17 @@ p {
     margin-block-end: 0;
 }
 
+li {
+    list-style-type: none;
+}
+
+ul {
+    margin-block-start: 0;
+}
+
 button {
     cursor: pointer;
+    border-width: 0;
 }
 
 button:disabled {
