+++ src/styles/components/_radiobutton.scss
@@ -1,0 +1,22 @@
+.container {
+    display: flex; 
+    justify-content: center; 
+    align-items: center; 
+}
+
+.radiobutton { 
+    padding: 18px;
+    border-top: 2px solid $dark-blue;
+    background: $off-white; 
+    display: flex; 
+    justify-content: left; 
+    width: 280px; 
+    transition: "background-color 0.3s, color 0.3s"; 
+    font-size: $font-size-medium;
+}
+
+.selected { 
+    background: $dark-blue;
+    color: $off-white; 
+    border-color: $dark-blue; 
+}
