+++ src/styles/styles.scss
@@ -4,3 +4,4 @@
 @import './components/sort';
 @import './components/button';
 @import './components/hotel';
+@import './components/radiobutton';
